Added authorization

+++ backend/routes/posts.js
@@ -31,7 +31,8 @@ router.post('', checkAuth,multer({storage: storage}).single('image') ,(req, res,
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -54,11 +55,15 @@ router.put('/:id', checkAuth, multer({storage: storage}).single('image'), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: 'Updated Successfully'});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: 'Updated Successfully'});
+    } else {
+      res.status(401).json({message: 'Unauthorized User'});
+    }
   });
 });
 
@@ -94,8 +99,12 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then((result) => {
-    res.status(200).json({message: 'Post deleted'});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then((result) => {
+    if (result.n > 0) {
+      res.status(200).json({message: 'Post deleted'});
+    } else {
+      res.status(401).json({message: 'Unauthorized User'});
+    }
   });
 });
